+++ src/main/java/me/kelsi/events/tabEvent.java
@@ -1,7 +1,5 @@
 package me.kelsi.events;
 
-import com.comphenix.protocol.ProtocolLibrary;
-import com.comphenix.protocol.ProtocolManager;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.Bukkit;
@@ -15,12 +13,16 @@ import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.plugin.java.JavaPlugin;
 import me.kelsi.utils.color;
 
-import java.util.List;
-
 public class tabEvent implements Listener {
 
 
     JavaPlugin plugin; public tabEvent(JavaPlugin plugin) {this.plugin = plugin;}
+
+    FileConfiguration config = plugin.getConfig();
+    String tps = String.valueOf(Bukkit.getTPS());
+    String online = String.valueOf(Bukkit.getOnlinePlayers().size());
+    String header_text = color.color(config.getString("tab.header.a") + color.color(config.getString("tab.header.b")));
+    String footer_text = color.color(config.getString("tab.footer.a").replace("value", online) + color.color(config.getString("tab.footer.b").replace("value", tps)));
 
     @EventHandler
     public void onPlayerJoin(PlayerJoinEvent event) {
@@ -28,6 +30,11 @@ public class tabEvent implements Listener {
         Player player = event.getPlayer();
         if (plugin.getConfig().getBoolean("change.pt")) {
             plugin.getConfig().set("player_timer", player.getName());
+        }
+        for (Player p : Bukkit.getOnlinePlayers()) {
+            BaseComponent[] header = TextComponent.fromLegacyText(color.color(header_text));
+            BaseComponent[] footer = TextComponent.fromLegacyText(color.color(footer_text));
+            p.setPlayerListHeaderFooter(header, footer);
         }
 
     }
@@ -39,6 +46,11 @@ public class tabEvent implements Listener {
         if (plugin.getConfig().getBoolean("change.pt")) {
             plugin.getConfig().set("player_timer", player.getName());
         }
+        for (Player p : Bukkit.getOnlinePlayers()) {
+            BaseComponent[] header = TextComponent.fromLegacyText(color.color(header_text));
+            BaseComponent[] footer = TextComponent.fromLegacyText(color.color(footer_text));
+            p.setPlayerListHeaderFooter(header, footer);
+        }
 
     }
 
@@ -46,30 +58,16 @@ public class tabEvent implements Listener {
     public void onTabUpdate(PlayerInteractEvent event) {
         Player player = event.getPlayer();
         String timer_name = plugin.getConfig().getString("player_timer");
-        ProtocolManager manager = ProtocolLibrary.getProtocolManager();
-        FileConfiguration config = plugin.getConfig();
 
-        String header_text = color.color(config.getString("tab.header.a") + color.color(config.getString("tab.header.b")));
-        String footer_text = color.color(config.getString("tab.footer.a") + color.color(config.getString("tab.footer.b")));
 
         if (player.getName().equalsIgnoreCase(timer_name)) {
 
-            /*PacketContainer pc = manager.createPacket((PacketType.Play.Server.PLAYER_LIST_HEADER_FOOTER));
-
-            pc.getChatComponents().write(0, WrappedChatComponent.fromText(color.color(plugin.getConfig().getString("tab.header", ""))))
-                    .write(1, WrappedChatComponent.fromText(color.color(plugin.getConfig().getString("tab, footer"))));
-            try {
-                for (Player p : Bukkit.getOnlinePlayers()) {
-                    manager.sendServerPacket(p, pc);
-                }
-            } catch (Exception ex) {
-                ex.printStackTrace();
-            }*/
             for (Player p : Bukkit.getOnlinePlayers()) {
                 BaseComponent[] header = TextComponent.fromLegacyText(color.color(header_text));
                 BaseComponent[] footer = TextComponent.fromLegacyText(color.color(footer_text));
                 p.setPlayerListHeaderFooter(header, footer);
             }
+
         }
     }
 }