+++ CHANGELOG.md
@@ -12,3 +12,9 @@
 # v. 1.2.2
 ### Added:
 - tab fix
+
+# v. 1.2.3
+### Added:
+- finaly tab fix
+- fix played hours statistic in /stats player <player>
+- optimize plugin code